+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The following changes have been implemented but not released yet:
 
 ### New features
 
-- `@inrupt/generate-oidc-token` can now be used with `npx`
+- `@inrupt/generate-oidc-token` can now be used with `npx`.
+- An interactive prompt helps the user provide the relevant options.
 
 The following sections document changes that have been released already: